Remove ambiguities in timing namespace and move RCTree
First, when using Arc as the entity class name of the TimingArc
Entity System, it caused an ambiguity with the class TimingGraph,
where it needs the use of the name Arc (directed edge). Another
modification was to place enums and class Arc within the scope
of the TimingGraph class, thus giving full responsibility for
using it.
Second, was moved RCTree from timing namespace to interconnection
because it is more suitable for such class.