Create SFML targets correctly.
Previously it was tried to create a target containing all the modu-
les of the library but cmake does not create targets of libraries
that uses a list of files containing all the modules. The solution
was to create each module as an independent target and so use them
in the gui.
Also made the appropriate changes in gitlab CI.