Group nodes and arcs data into a single class
Previously, to access the time data of a circuit in the calcula-
tion of the generic STA, a structure was used that grouped data
of the nodes and data of the arcs, but they were separated clas-
ses. Given that the two will always be created together, it has
changed the responsibility of maintaining and managing all of this
data to the TimingData, making it a class of fact rather than an
intermediate structure.